#18 renaming component s folder
